% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{get_bananas}
\alias{get_bananas}
\alias{report_bananas}
\alias{list_bananas}
\alias{ping_bananas}
\title{Reports}
\usage{
get_bananas(id = NULL, banana = NULL, color = c("reports", "lists"), ...)

report_bananas(id = NULL, banana = NULL, ...)

list_bananas(id = NULL, banana = NULL, ...)

ping_bananas()
}
\arguments{
\item{id}{Id of banana to fetch.}

\item{banana}{Banana type to fetch.}

\item{color}{Color of banana.}

\item{...}{Banana parameters.}
}
\value{
A \code{list} of bananas.
}
\description{
Get bananas on reports!
}
\examples{
\dontrun{
set_bananas("xXXXXXxxXXXxx")

ping_bananas() # healthy bananas

bananas <- report_bananas() # get reports
bananas2 <- get_bananas() # get reports
identical(bananas, bananas2)

list_of_bananas <- list_bananas()
}

}
