% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_bananas}
\alias{report_bananas}
\title{Reports}
\usage{
report_bananas(id = NULL, banana = NULL)
}
\arguments{
\item{id}{Id of banana to fetch.}

\item{banana}{Banana type to fetch.}
}
\value{
A \code{list} of bananas.
}
\description{
Get bananas on reports!
}
\examples{
\dontrun{
set_bananas("xXXXXXxxXXXxx")

reports <- report_bananas()
}

}
