% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{set_bananas}
\alias{set_bananas}
\title{Setup}
\usage{
set_bananas(key, server = NULL, version = "3.0")
}
\arguments{
\item{key}{Your API key.}

\item{server}{Part of the URL corresponds to the data center for your account, see details.}

\item{version}{API version.}
}
\description{
Set your bananas.
}
\details{
\code{server}: The last part of your MailChimp API key, and/or first part of your mailchimp url.
}
\examples{
\dontrun{set_bananas("xXXXXXxxXXXxx")}

}
