% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_fruits_I_can_eat.R
\name{how_many_fruits_I_can_eat}
\alias{how_many_fruits_I_can_eat}
\title{A function that tells you how many fruits you can eat, and which types}
\usage{
how_many_fruits_I_can_eat(apples, oranges, fruit_counts)
}
\arguments{
\item{oranges}{types of oranges}

\item{fruit_counts}{amount of fruit you can eat}

\item{apple}{types of apples}
}
\description{
A function that tells you how many fruits you can eat, and which types
}
\examples{
how_many_fruits_I_can_eat(4,"Fuji", "cuties")
}
